api/requirements.txt
uvicorn
fastapi
motor
starlette
loguru
python-dotenv
bcrypt
python-jose